separated sponsors in tiers
