+++ src/App.css
@@ -22,6 +22,7 @@ h1 {
 h2 {
   font-weight: 600;
   font-size: 30px;
+  text-align: center;
 }
 
 h6 {
